requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backoff==2.2.1
base58==2.1.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
bitarray==2.6.1
bleach==5.0.1
cached-property==1.5.2
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
cytoolz==0.12.1
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.4
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-redis==5.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
entrypoints==0.4
eth-abi==2.2.0
eth-account==0.5.9
eth-bloom==1.0.4
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-tester==0.6.0b7
eth-typing==2.3.0
eth-utils==1.9.5
executing==1.2.0
fastjsonschema==2.16.2
flower==1.2.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
futureproof==0.3.1
gql==3.4.0
graphql-core==3.2.3
hexbytes==0.2.3
humanize==4.4.0
idna==3.4
ipfshttpclient==0.8.0a2
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kombu==5.2.4
lru-dict==1.1.8
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
multiaddr==0.0.9
multidict==6.0.4
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
netaddr==0.8.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
plotly==5.11.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.19.5
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-ecc==5.2.0
py-evm==0.5.0a3
py-geth==3.10.0
pyarrow==10.0.1
pycparser==2.21
pycryptodome==3.16.0
pyethash==0.1.27
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
pysha3==1.0.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
redis==4.4.0
requests==2.28.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rlp==2.0.1
semantic-version==2.10.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
tenacity==8.1.0
terminado==0.17.1
tinycss2==1.2.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
trie==2.0.0a5
typing-extensions==3.10.0.2
tzdata==2022.7
ujson==5.6.0
uri-template==1.2.0
urllib3==1.26.13
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.31.3
web3-ethereum-defi==0.12
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
websockets==9.1
widgetsnbextension==4.0.5
yarl==1.8.2
